requirements.txt
arrow==0.15.5
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
config-parser==0.0.1
croniter==0.3.31
dateutils==0.6.7
decorator==4.4.1
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
Flask==1.1.1
httpie==2.0.0
idna==2.8
ipython==7.12.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
Markdown==3.2
MarkupSafe==1.1.1
mysqlclient==1.4.5
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.22.0
rq==1.2.2
rq-dashboard==0.6.1
rq-scheduler==0.9.1
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
